+++ TeamProject/static/js/board/post/comment/render.js
@@ -1,7 +1,7 @@
-import * as MAIN from "../main.js";
+import * as MAIN from "../../main.js";
 import * as EVENT from "./event.js";
 import * as INDEX from "./index.js";
-import * as FETCH from "../fetch.js";
+import * as FETCH from "../../fetch.js";
 
 export function post_comment_list(comment, user_data, login_currentUserData) {
     let comment_html = `<div class = "comment_item" id="comment_id_${comment.id}"><div class="comment_top">` +
@@ -30,13 +30,11 @@ export function post_comment_list(comment, user_data, login_currentUserData) {
   export async function post_comment(comments) {
     let text = '';
     const login_currentUserData = await FETCH.fetch_userinfo();
-  
     for (let i = comments.length - 1; i >= 0; i--) {
       const user_data = await FETCH.fetch_getUserdata(comments[i].userid);
       text += post_comment_list(comments[i], user_data, login_currentUserData);
       //수정
     }
-    
     document.querySelector('.comment_list').innerHTML = text;
     EVENT.add_comment_likes();
     EVENT.add_comment_report();
@@ -53,6 +51,6 @@ export function post_comment_list(comment, user_data, login_currentUserData) {
     ele.replaceChild(ele_textarea, ele.childNodes[1]);
     const button = ele.querySelector(`#updateComment__${id}`).parentNode;
     const new_button = await MAIN.get_htmlObject('input',
-      ['type', 'id', 'value'], ['button',`updateComment__${id}`, '완료']);
+      ['type', 'id', 'value'], ['button',`updateCommentSubmit__${id}`, '완료']);
     button.replaceChild(new_button, button.childNodes[0]);
   }
