+++ TeamProject/static/js/board/post/fetch.js
@@ -147,28 +147,3 @@ export async function insert_post_report(id) {
 
     }
 }
-
-function check_token(token) {
-    if (token == null) {
-        alert('로그인을 먼저 해주세요.');
-        return false;
-    } else return true;
-}
-
-function check_response_json(response) {
-    if(response.ok) {
-        
-    } else {
-        console.log("HTTP-ERROR : " + response.status);
-        return response.status;
-    }
-}
-
-function check_response_boolean(response) {
-    if (response.ok){
-        return true;
-    } else {
-        console.log("HTTP-ERROR : " + response.status);
-        return response.status;
-    }
-}
